# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GroverRunRequest": {
        "properties": {
          "candidates": {
            "default": 4,
            "enum": [
              4,
              8,
              16
            ],
            "title": "Candidates",
            "type": "integer"
          },
          "iterations": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Iterations",
            "type": "integer"
          },
          "preparation": {
            "default": "uniform",
            "enum": [
              "uniform",
              "missing"
            ],
            "title": "Preparation",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "shots": {
            "default": 256,
            "minimum": 1.0,
            "title": "Shots",
            "type": "integer"
          },
          "target": {
            "default": 2,
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Target",
            "type": "integer"
          }
        },
        "title": "GroverRunRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend for the Quantum Learning Laboratory.\n\nThe quantum engine computes, the authored evaluator grades, and the tutor\nteaches. Private rubrics and answer keys are never returned by any route in\nthis contract.",
    "title": "Quantum Learning Laboratory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/grover/analysis": {
      "get": {
        "description": "Analytical prediction curve. Labelled as a formula evaluation, not a run.",
        "operationId": "grover_analysis_api_v1_grover_analysis_get",
        "parameters": [
          {
            "in": "query",
            "name": "candidates",
            "required": false,
            "schema": {
              "default": 4,
              "title": "Candidates",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grover Analysis Api V1 Grover Analysis Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grover Analysis",
        "tags": [
          "previews"
        ]
      }
    },
    "/api/v1/grover/classical-trace": {
      "get": {
        "operationId": "grover_classical_trace_api_v1_grover_classical_trace_get",
        "parameters": [
          {
            "in": "query",
            "name": "candidates",
            "required": false,
            "schema": {
              "default": 8,
              "title": "Candidates",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": 5,
              "title": "Target",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grover Classical Trace Api V1 Grover Classical Trace Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grover Classical Trace",
        "tags": [
          "previews"
        ]
      }
    },
    "/api/v1/grover/runs": {
      "post": {
        "description": "Run the bounded Grover preset through the ordinary engine.",
        "operationId": "grover_run_api_v1_grover_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroverRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grover Run Api V1 Grover Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grover Run",
        "tags": [
          "previews"
        ]
      }
    },
...
    "/api/v1/hints": {
      "post": {
        "description": "Authored hints. Delivery is recorded so assisted success stays distinguishable.",
        "operationId": "request_hint_api_v1_hints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Request Hint Api V1 Hints Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Hint",
        "tags": [
          "hints"
        ]
      }
    },
...
    "/api/v1/previews/{algorithm_id}/encounters": {
      "post": {
        "description": "Record Seen for a preview. Seeing a preview never grants a higher level.",
        "operationId": "record_encounter_api_v1_previews__algorithm_id__encounters_post",
        "parameters": [
          {
            "in": "path",
            "name": "algorithm_id",
            "required": true,
            "schema": {
              "title": "Algorithm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewEncounterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Encounter Api V1 Previews  Algorithm Id  Encounters Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Encounter",
        "tags": [
          "previews"
        ]
      }
    },
...
    "/api/v1/revisions": {
      "post": {
        "description": "Commit an edit. Every committed edit becomes a new immutable revision.",
        "operationId": "create_revision_api_v1_revisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Revision Api V1 Revisions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Revision",
        "tags": [
          "lab"
        ]
      }
    },
    "/api/v1/runs": {
      "post": {
        "description": "Compute a new run. Each shot restarts the declared preparation.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Run Api V1 Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "lab"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "description": "Replay reads a saved run, including the outcomes it actually recorded.",
        "operationId": "read_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Run Api V1 Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Run",
        "tags": [
          "lab"
        ]
      }
    },
    "/api/v1/runs/{run_id}/repeat-measurement": {
      "post": {
        "description": "Re-read the recorded conditional trajectory. Not a fresh shot.",
        "operationId": "repeat_measurement_api_v1_runs__run_id__repeat_measurement_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Repeat Measurement Api V1 Runs  Run Id  Repeat Measurement Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repeat Measurement",
        "tags": [
          "lab"
        ]
      }
    },
...
  }
}